@@ -19,17 +19,18 @@ setuptools.setup(
         "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
         "Operating System :: OS Independent",
     ],
-  python_requires='>=3.8',
-  install_requires=['scipy==1.5.4',
-                    'numpy==1.21.6',
-                    'pandas==1.1.5',
-                    'matplotlib==3.3.3',
-                    'h5py==3.1.0',
-                    'PySide2==5.15.2',
-                    'pyqtgraph==0.11.0',
-                    'numba==0.52.0',
-                    'pyopengl==3.1.1a1',
-                    'pyopengl-accelerate==3.1.3b1',
+  python_requires='>=3.11',
+  install_requires=['scipy==1.11.3',
+                    'numpy==1.26.1',
+                    'pandas==2.1.2',
+                    'matplotlib==3.8.1',
+                    'h5py==3.10.0',
+                    'PySide6==6.6.2',
+                    'pyqtgraph==0.13.3',
+                    'numba==0.58.1',
+                    'pyopengl==3.1.7',
+                    'pyopengl-accelerate==3.1.7',
+                    'pyEDFlib==0.1.36',
                     'rsa',
                     'jupyter',
                     ],
