@@ -6,7 +6,7 @@ import glob, os
 from datetime import datetime
 from pyecog2.annotations_module import AnnotationPage
 from scipy import signal
-from PySide2 import QtCore
+from PySide6 import QtCore
 import pyqtgraph as pg
 from timeit import default_timer as timer
 from pyedflib import EdfReader
@@ -269,6 +269,8 @@ class FileBuffer():  # Consider translating this to cython
         self.eeg_init_time = eeg_init_time
         self.eeg_duration = eeg_duration
         self.verbose = verbose
+        self.montage = np.eye(1)
+        self.apply_montage = False
 
     def add_file_to_buffer(self, fname):
         if fname in self.files:  # skip if file is already buffered
@@ -357,7 +359,6 @@ class FileBuffer():  # Consider translating this to cython
     def get_t_max_for_live_plot(self):
         return max([r[1] for r in self.data_ranges])
 
-
     def get_data_from_range(self, trange, channel=None, n_envelope=None, for_plot=False, filter_settings=(False,0,0)):
         # First check if data is already buffered, most of the time this will be the case:
         if trange[0] >= self.range[0] and trange[1] <= self.range[1]:
@@ -405,7 +406,14 @@ class FileBuffer():  # Consider translating this to cython
         enveloped_data = []
         enveloped_time = []
         no_downsampling = True
+
         for i, data in enumerate(self.data):
+            if self.montage.shape[1] != data.shape[1]:
+                if (self.montage == np.eye(self.montage.shape[1])).all(): # create apropriately sized matrix if it is Identity
+                    self.montage = np.eye(data.shape[1])
+                else:
+                    print('Montage shape incompatible with data shape')
+                    raise IndexError
 
             if channel is not None and channel>=data.shape[1]:
                 continue # skip this file becuase it does not have channel with required index
@@ -419,15 +427,32 @@ class FileBuffer():  # Consider translating this to cython
             # print('Downsampling ratio:', ds,file_envlopes,sample_ranges)
             if channel is None:
                 # poor coding here, we are not computing proper envelopes, but it'll do for now because this is only
-                # used for coputing channel scallings so far
-                enveloped_data.append(dV*data[start:stop:ds, :])
+                # used for computing channel scallings so far
+                if self.apply_montage and not (self.montage==np.eye(self.montage.shape[0])).all(): # avoid multiplication if montage is identity matrix
+                    enveloped_data.append(data[start:stop:ds, :]@(dV*self.montage.T))
+                else:
+                    enveloped_data.append(data[start:stop:ds, :]*dV)
                 if ds != 0:
                     no_downsampling = False
             elif ds == 1:
                 # Small enough to display with no intervention.
-                enveloped_data.append(dV*data[start:stop, channel].reshape(-1, 1))
+                if self.apply_montage:
+                    montage_channel = self.montage[channel,:]
+                    nzmi = np.nonzero(montage_channel)  # non zero indices of montage to avoid unecessary computations
+                else:
+                    montage_channel = np.zeros(data.shape[1])  # no montage is applied
+                    montage_channel[channel] = 1
+                    nzmi = np.array([channel])
+                enveloped_data.append((data[start:stop, nzmi]@(montage_channel[nzmi].T*dV)).reshape(-1, 1))
             else:
                 no_downsampling = False
+                if self.apply_montage:
+                    montage_channel = self.montage[channel, :]
+                    nzmi = np.nonzero(montage_channel)  # non zero indices of montage to avoid unecessary computations
+                else:
+                    montage_channel = np.zeros(data.shape[1])  # no montage is applied
+                    montage_channel[channel] = 1
+                    nzmi = np.array([channel])
                 # Here convert data into a down-sampled array suitable for visualizing.
                 # Must do this piecewise to limit memory usage.
                 dss = 1
@@ -446,7 +471,7 @@ class FileBuffer():  # Consider translating this to cython
                     # read data in chunks of ~1M samples
                     chunkSize = int((1e6 // ds) * ds)
                     while sourcePtr < stop - 1:
-                        chunk_data = data[sourcePtr:min(stop, sourcePtr + chunkSize):dss, channel]
+                        chunk_data = data[sourcePtr:min(stop, sourcePtr + chunkSize):dss, nzmi]@montage_channel[nzmi].T
                         sourcePtr += chunkSize
                         # reshape chunk to be integer multiple of ds
                         chunk_data = chunk_data[:(len(chunk_data) // ds) * ds].reshape(len(chunk_data) // ds, ds)
@@ -468,8 +493,7 @@ class FileBuffer():  # Consider translating this to cython
 
             enveloped_time.append(np.linspace(start / fs + self.data_ranges[i][0], (stop-1) / fs + self.data_ranges[i][0],
                                               len(enveloped_data[-1])).reshape(-1, 1))
-            # enveloped_time.append(np.linspace(start / fs + self.data_ranges[i][0], (stop-1) / fs + self.data_ranges[i][0],
-            #                                   len(enveloped_data[-1])).reshape(-1, 1))
+
             if len(enveloped_time[-1]) == 0:
                 del (enveloped_time[-1])
                 del (enveloped_data[-1])
@@ -493,21 +517,21 @@ class FileBuffer():  # Consider translating this to cython
         if len(enveloped_data) > 0:
             data = np.vstack(enveloped_data)
             time = np.vstack(enveloped_time)
-            # if for_plot and filter_settings[0]: # apply LP filter only for plots
-            #     fs = 2/(time[2]-time[0])
-            #     nyq = 0.5 * fs[0]
-            #     hpcutoff = min(max(filter_settings[1] / nyq, 0.001), .5)
-            #     data = data - np.mean(data)
-            #     lpcutoff = min(max(filter_settings[2] / nyq, 0.001), 1)
-            #     # for some reason the bandpass butterworth filter is very unstable
-            #     if lpcutoff<.99:  # don't apply filter if LP cutoff freqquency is above nyquist freq.
-            #         # if self.verbose: print('applying LP filter to display data:', filter_settings, fs, nyq, lpcutoff)
-            #         b, a = signal.butter(2, lpcutoff, 'lowpass', analog=False)
-            #         data = signal.filtfilt(b, a, data,axis =0,method='gust')
-            #     if hpcutoff > .001: # don't apply filter if HP cutoff frequency too low.
-            #         # if self.verbose: print('applying HP filter to display data:', filter_settings, fs, nyq, hpcutoff)
-            #         b, a = signal.butter(2, hpcutoff, 'highpass', analog=False)
-            #         data = signal.filtfilt(b, a, data,axis =0,method='gust')
+            if for_plot and filter_settings[0]: # apply LP filter only for plots
+                fs = 2/(time[2]-time[0])
+                nyq = 0.5 * fs[0]
+                hpcutoff = min(max(filter_settings[1] / nyq, 0.001), .5)
+                data = data - np.mean(data)
+                lpcutoff = min(max(filter_settings[2] / nyq, 0.001), 1)
+                # for some reason the bandpass butterworth filter is very unstable
+                if lpcutoff<.99:  # don't apply filter if LP cutoff freqquency is above nyquist freq.
+                    # if self.verbose: print('applying LP filter to display data:', filter_settings, fs, nyq, lpcutoff)
+                    b, a = signal.butter(2, lpcutoff, 'lowpass', analog=False)
+                    data = signal.filtfilt(b, a, data,axis =0,method='gust')
+                if hpcutoff > .001: # don't apply filter if HP cutoff frequency too low.
+                    # if self.verbose: print('applying HP filter to display data:', filter_settings, fs, nyq, hpcutoff)
+                    b, a = signal.butter(2, hpcutoff, 'highpass', analog=False)
+                    data = signal.filtfilt(b, a, data,axis =0,method='gust')
         else:
             data = np.zeros((0,1))
             time = np.zeros((0,1))
@@ -554,7 +578,12 @@ class Project():
         self.current_animal.annotations.copy_from(self.main_model.annotations,connect_history=False,quiet=True)
         self.main_model.annotations.copy_from(animal.annotations,quiet=True)
         self.current_animal = animal
+        apply_montage = None
+        if hasattr(self,'file_buffer') and hasattr(self.file_buffer,'file_buffer'):
+            apply_montage, montage = self.file_buffer.apply_montage, self.file_buffer.montage # keep previous montage settings - poor coding practice
         self.file_buffer = FileBuffer(self.current_animal)
+        if apply_montage is not None:
+            self.file_buffer.apply_montage, self.file_buffer.montage = apply_montage, montage
         self.main_model.annotations.sigLabelsChanged.emit('')
         logger.info(f'ProjectClass set_current_animal ran in {timer()-start_t} seconds')
 
@@ -604,6 +633,7 @@ class Project():
 
         if not hasattr(self,'filter_settings'):  #Backwards compatibility
             self.filter_settings = (False, 0, 1e6)
+
         self.main_model.sigProjectChanged.emit()
         return (new_dirname, orig_dirname)
 
