@@ -93,9 +93,9 @@ t.setParameters(p, showTop=False)
 t.setWindowTitle('pyqtgraph example: Parameter Tree')
 
 win = QtGui.QWidget()
-layout = QtGui.QGridLayout()
+layout = QGridLayout()
 win.setLayout(layout)
-layout.addWidget(QtGui.QLabel("These are two views of the same data. They should always display the same values."), 0,
+layout.addWidget(QLabel("These are two views of the same data. They should always display the same values."), 0,
                  0, 1, 1)
 layout.addWidget(t, 1, 0, 1, 1)
 win.show()
@@ -106,4 +106,4 @@ if __name__ == '__main__':
     import sys
 
     if (sys.flags.interactive != 1) or not hasattr(QtCore, 'PYQT_VERSION'):
-        QtGui.QApplication.instance().exec_()
+        QtGui.QApplication.instance().exec()
