@@ -733,8 +733,8 @@ class DataHandler:
 
 
 import multiprocessing as mp
-from PySide2 import QtCore
-from PySide2.QtWidgets import QMenuBar, QGridLayout, QApplication, QWidget, QPlainTextEdit, QMainWindow
+from PySide6 import QtCore
+from PySide6.QtWidgets import QMenuBar, QGridLayout, QApplication, QWidget, QPlainTextEdit, QMainWindow
 import sys
 
 class GuiMain(QMainWindow):
@@ -757,10 +757,10 @@ class OpenWindowProcess(mp.Process):
         app = QApplication(sys.argv)
         window = QMainWindow()
         window.show()
-        sys.exit(app.exec_())
+        sys.exit(app.exec())
 
 
 if __name__ == '__main__':
     app = QApplication(sys.argv)
     application = GuiMain()
-    sys.exit(app.exec_())
+    sys.exit(app.exec())
