@@ -1,6 +1,6 @@
 import sys, os
-from PySide2 import QtGui, QtCore
-from PySide2.QtWidgets import QGridLayout, QApplication, QWidget, QMainWindow, QTextBrowser, QPushButton, QFileDialog, \
+from PySide6 import QtGui, QtCore
+from PySide6.QtWidgets import QGridLayout, QApplication, QWidget, QMainWindow, QTextBrowser, QPushButton, QFileDialog, \
     QProgressBar, QCheckBox
 
 from pyqtgraph.parametertree import Parameter
@@ -258,4 +258,4 @@ if __name__ == '__main__':
     window = FeatureExtractorWindow()
     window.setGeometry(500, 300, 300, 200)
     window.show()
-    sys.exit(app.exec_())
+    sys.exit(app.exec())
