@@ -1,7 +1,7 @@
 import sys, os, glob
 from datetime import datetime
-from PySide2 import QtGui, QtCore
-from PySide2.QtWidgets import QGridLayout, QApplication, QWidget, QMainWindow, QTextBrowser, QPushButton, QFileDialog
+from PySide6 import QtGui, QtCore
+from PySide6.QtWidgets import QGridLayout, QApplication, QWidget, QMainWindow, QTextBrowser, QPushButton, QFileDialog
 
 from pyqtgraph.parametertree import Parameter
 from pyecog2.ndf_converter import NdfFile, DataHandler
@@ -150,7 +150,7 @@ class NDFConverterWindow(QMainWindow):
         dialog = QFileDialog(self)
         dialog.setDirectory(self.folder2convert)
         dialog.setWindowTitle('Select NDF directory')
-        dialog.setFileMode(QFileDialog.DirectoryOnly)
+        dialog.setFileMode(QFileDialog.Directory)
         # dialog.setOption(QFileDialog.DontUseNativeDialog, True)
         dialog.setAcceptMode(QFileDialog.AcceptOpen)
         if dialog.exec():
@@ -196,7 +196,7 @@ class NDFConverterWindow(QMainWindow):
         dialog = QFileDialog(self)
         dialog.setDirectory(self.destination_folder)
         dialog.setWindowTitle('Select Destination directory')
-        dialog.setFileMode(QFileDialog.DirectoryOnly)
+        dialog.setFileMode(QFileDialog.Directory)
         # dialog.setOption(QFileDialog.DontUseNativeDialog, True)
         dialog.setAcceptMode(QFileDialog.AcceptOpen)
         if dialog.exec():
@@ -267,4 +267,4 @@ if __name__ == '__main__':
     window = NDFConverterWindow()
     window.setGeometry(500, 300, 300, 200)
     window.show()
-    sys.exit(app.exec_())
+    sys.exit(app.exec())
