@@ -1,6 +1,6 @@
 import sys
-from PySide2 import QtGui, QtCore
-from PySide2.QtWidgets import QGridLayout, QApplication, QWidget, QMainWindow, QTextBrowser, QPushButton, QFileDialog
+from PySide6 import QtGui, QtCore
+from PySide6.QtWidgets import QGridLayout, QApplication, QWidget, QMainWindow, QTextBrowser, QPushButton, QFileDialog
 
 from pyecog2.ProjectClass import Animal
 # from ..ProjectClass import Animal
@@ -160,7 +160,7 @@ class ProjectEditWindow(QMainWindow):
     def selectEEGFolder(self):
         dialog = QFileDialog(self)
         dialog.setWindowTitle('Select EEG directory')
-        dialog.setFileMode(QFileDialog.DirectoryOnly)
+        dialog.setFileMode(QFileDialog.Directory)
         # dialog.setOption(QFileDialog.DontUseNativeDialog, True)
         dialog.setAcceptMode(QFileDialog.AcceptOpen)
         if dialog.exec():
@@ -175,7 +175,7 @@ class ProjectEditWindow(QMainWindow):
     def selectVideoFolder(self):
         dialog = QFileDialog(self)
         dialog.setWindowTitle('Select EEG directory')
-        dialog.setFileMode(QFileDialog.DirectoryOnly)
+        dialog.setFileMode(QFileDialog.Directory)
         # dialog.setOption(QFileDialog.DontUseNativeDialog, True)
         dialog.setAcceptMode(QFileDialog.AcceptOpen)
         if dialog.exec():
@@ -244,4 +244,4 @@ if __name__ == '__main__':  # OBSOLETE ???
     window = ProjectEditWindow()
     window.setGeometry(500, 300, 300, 200)
     window.show()
-    sys.exit(app.exec_())
+    sys.exit(app.exec())
