@@ -4,9 +4,9 @@ Wavelet widget for EEG signals in pyecog
 """
 
 import pyqtgraph as pg
-from PySide2 import QtCore, QtGui
-from PySide2.QtWidgets import QApplication, QWidget
-from PySide2.QtCore import QRunnable, Slot, QThreadPool
+from PySide6 import QtCore
+from PySide6.QtWidgets import QApplication, QWidget, QGridLayout, QLabel
+from PySide6.QtCore import QRunnable, Slot, QThreadPool
 import numpy as np
 import scipy.signal as sg
 from timeit import default_timer as timer
@@ -403,9 +403,9 @@ class WaveletWindowItem(pg.GraphicsLayoutWidget):
                 # self.hist.setLevels(*self.hist_levels_cross)
 
         else:  # plotting normal wavelet
-            print('self.last_plot_was_wave:', self.last_plot_was_wave)
+            # print('self.last_plot_was_wave:', self.last_plot_was_wave)
             if not self.last_plot_was_wave:
-                print('reseting colormap')
+                # print('reseting colormap')
                 self.hist.gradient.loadPreset('viridis')
             self.last_plot_was_wave = True
             self.last_plot_was_cross = False
@@ -416,8 +416,9 @@ class WaveletWindowItem(pg.GraphicsLayoutWidget):
         self.img.resetTransform()
         ymin = np.log10(vf[0])
         ymax = np.log10(vf[-1])
-        self.img.translate(0,ymin)
-        self.img.scale(self.dt,(ymax-ymin)/self.data.shape[0])
+        # self.img.translate(0,ymin)
+        # self.img.scale(self.dt,(ymax-ymin)/self.data.shape[0])
+        self.img.setRect(0,ymin,self.dt*self.data.shape[1],(ymax-ymin))
         self.vb.setLimits(xMin=0, xMax=self.data.shape[1]*self.dt, yMin=ymin, yMax=ymax)
         self.vb.setRange(xRange=[0,self.data.shape[1]*self.dt])
         self.p1.setLabel('bottom', 'Time', units='s')
@@ -443,12 +444,12 @@ class WaveletWindowItem(pg.GraphicsLayoutWidget):
 class WaveletWindow(QWidget):
     def __init__(self, main_model = None):
         QWidget.__init__(self)
-        self.layout = QtGui.QGridLayout()
+        self.layout = QGridLayout()
         self.setLayout(self.layout)
         self.wavelet_item = WaveletWindowItem(main_model)
 
         self.controls_widget = QWidget()
-        self.controls_layout = QtGui.QGridLayout()
+        self.controls_layout = QGridLayout()
         self.controls_widget.setLayout(self.controls_layout)
         self.channel_spin = pg.SpinBox(value=0,bounds=[0,None], int=True, minStep=1, step=1,compactHeight=False)
         self.channel_spin.valueChanged.connect(self.wavelet_item.setChannel)
@@ -456,15 +457,18 @@ class WaveletWindow(QWidget):
         self.cross_channel_spin.valueChanged.connect(self.wavelet_item.setCrossChannel)
         self.R_spin = pg.SpinBox(value=14.0, bounds=[5, None],step=1,compactHeight=False)
         self.R_spin.valueChanged.connect(self.wavelet_item.setR)
-        self.controls_layout.addWidget(QtGui.QLabel('Channel'),0,0)
+        self.controls_layout.addWidget(QLabel('Channel'),0,0)
         self.controls_layout.addWidget(self.channel_spin,0,1,)
-        self.controls_layout.addWidget(QtGui.QLabel('Wavelet factor R'),0,2)
+        self.controls_layout.addWidget(QLabel('Wavelet factor R'),0,2)
         self.controls_layout.addWidget(self.R_spin,0,3)
-        self.controls_layout.addWidget(QtGui.QLabel('Cross wavelet Channel'),0,4)
+        self.controls_layout.addWidget(QLabel('Cross wavelet Channel'),0,4)
         self.controls_layout.addWidget(self.cross_channel_spin,0,5)
 
         self.layout.addWidget(self.controls_widget,1,0)
         self.layout.addWidget(self.wavelet_item,0,0)
+
+    def update_data(self):
+        self.wavelet_item.update_data()
 
 ## Start Qt event loop unless running in interactive mode or using pyside.
 if __name__ == '__main__':
@@ -473,7 +477,7 @@ if __name__ == '__main__':
     app = QApplication(sys.argv)
     w = WaveletWindow()
     w.show()
-    sys.exit(app.exec_())
+    sys.exit(app.exec())
 
 #
 # if __name__ == '__main__':
@@ -481,7 +485,7 @@ if __name__ == '__main__':
 #     player = VideoWindow()
 #     player.resize(640, 480)
 #     player.show()
-#     sys.exit(app.exec_())
+#     sys.exit(app.exec())
 
 class LogAxis(pg.AxisItem):
     def tickStrings(self, values, scale, spacing):
